16BEC0577_VL2019205006395_PE003.pdf
Augmented reality has created quite a stir and many consumers are wondering how this latest technology will actually have an impact on them as it becomes more widely used. For anyone that uses a mobile device for daily activities, augmented reality is the future that will allow consumers to experience a reality that is based on personal needs and desires. AR will present a completely new way to engage and will expand the abilities of retailers as well. The possibilities of augmented reality are literally endless and when combined with et technology of mobile devices, which is already quite powerful, AR will allow for geo-tracking that will allow amazing experiences for consumers.
Augmented reality will also have a strong impact on society. The marketing and advertisement fields will explode with augmented reality devices. The mobile applications that are being developed will offer facial recognition software that will mainstream quickly. For everyday commuters and drivers, navigation devices will be built into the cars and mobile devices. This is already in the works, with some car manufacturers working to implement augmented reality windshields that will help drivers navigate without taking their eyes off the road.
Considering this I developed an application for creating augmented reality content on dynamic surfaces i.e., surfaces which are moving and thus provide a means for the implementation of augmented reality in real life scenarios, where virtual objects can be successfully utilised for the day to day activities of any commoner.